+++ node/src/integration-test/kotlin/net/corda/services/messaging/MQSecurityAsNodeTest.kt
@@ -6,7 +6,6 @@ import net.corda.node.services.messaging.ArtemisMessagingComponent.Companion.RPC
 import net.corda.testing.messaging.SimpleMQClient
 import org.apache.activemq.artemis.api.config.ActiveMQDefaultConfiguration
 import org.apache.activemq.artemis.api.core.ActiveMQClusterSecurityException
-import org.apache.activemq.artemis.api.core.ActiveMQNotConnectedException
 import org.apache.activemq.artemis.api.core.ActiveMQSecurityException
 import org.assertj.core.api.Assertions.assertThatExceptionOfType
 import org.junit.Test
@@ -16,7 +15,7 @@ import org.junit.Test
  */
 class MQSecurityAsNodeTest : MQSecurityTest() {
     override fun createAttacker(): SimpleMQClient {
-        return clientTo(alice.configuration.messagingAddress)
+        return clientTo(alice.configuration.p2pAddress)
     }
 
     override fun startAttacker(attacker: SimpleMQClient) {
@@ -30,7 +29,7 @@ class MQSecurityAsNodeTest : MQSecurityTest() {
 
     @Test
     fun `only the node running the broker can login using the special node user`() {
-        val attacker = clientTo(alice.configuration.messagingAddress)
+        val attacker = clientTo(alice.configuration.p2pAddress)
         assertThatExceptionOfType(ActiveMQSecurityException::class.java).isThrownBy {
             attacker.start(NODE_USER, NODE_USER)
         }
@@ -38,7 +37,7 @@ class MQSecurityAsNodeTest : MQSecurityTest() {
 
     @Test
     fun `login as the default cluster user`() {
-        val attacker = clientTo(alice.configuration.messagingAddress)
+        val attacker = clientTo(alice.configuration.p2pAddress)
         assertThatExceptionOfType(ActiveMQClusterSecurityException::class.java).isThrownBy {
             attacker.start(ActiveMQDefaultConfiguration.getDefaultClusterUser(), ActiveMQDefaultConfiguration.getDefaultClusterPassword())
         }
@@ -46,7 +45,7 @@ class MQSecurityAsNodeTest : MQSecurityTest() {
 
     @Test
     fun `login without a username and password`() {
-        val attacker = clientTo(alice.configuration.messagingAddress)
+        val attacker = clientTo(alice.configuration.p2pAddress)
         assertThatExceptionOfType(ActiveMQSecurityException::class.java).isThrownBy {
             attacker.start()
         }
@@ -56,7 +55,7 @@ class MQSecurityAsNodeTest : MQSecurityTest() {
     fun `login to a non ssl port as a node user`() {
         val attacker = clientTo(alice.configuration.rpcAddress!!, sslConfiguration = null)
         assertThatExceptionOfType(ActiveMQSecurityException::class.java).isThrownBy {
-            attacker.start(NODE_USER, NODE_USER)
+            attacker.start(NODE_USER, NODE_USER, enableSSL = false)
         }
     }
 
@@ -64,7 +63,7 @@ class MQSecurityAsNodeTest : MQSecurityTest() {
     fun `login to a non ssl port as a peer user`() {
         val attacker = clientTo(alice.configuration.rpcAddress!!, sslConfiguration = null)
         assertThatExceptionOfType(ActiveMQSecurityException::class.java).isThrownBy {
-            attacker.start(PEER_USER, PEER_USER)  // Login as a peer
+            attacker.start(PEER_USER, PEER_USER, enableSSL = false)  // Login as a peer
         }
     }
 }