+++ node/src/main/kotlin/net/corda/node/driver/Driver.kt
@@ -342,16 +342,16 @@ open class DriverDSL(
 
     override fun startNode(providedName: String?, advertisedServices: Set<ServiceInfo>,
                            rpcUsers: List<User>, customOverrides: Map<String, Any?>): ListenableFuture<NodeHandle> {
-        val messagingAddress = portAllocation.nextHostAndPort()
+        val p2pAddress = portAllocation.nextHostAndPort()
         val rpcAddress = portAllocation.nextHostAndPort()
         val webAddress = portAllocation.nextHostAndPort()
         val debugPort = if (isDebug) debugPortAllocation.nextPort() else null
-        val name = providedName ?: "${pickA(name)}-${messagingAddress.port}"
+        val name = providedName ?: "${pickA(name)}-${p2pAddress.port}"
 
         val baseDirectory = driverDirectory / name
         val configOverrides = mapOf(
                 "myLegalName" to name,
-                "messagingAddress" to messagingAddress.toString(),
+                "p2pAddress" to p2pAddress.toString(),
                 "rpcAddress" to rpcAddress.toString(),
                 "webAddress" to webAddress.toString(),
                 "extraAdvertisedServiceIds" to advertisedServices.map { it.toString() },
@@ -382,7 +382,7 @@ open class DriverDSL(
         registerProcess(processFuture)
         return processFuture.flatMap { process ->
             // We continue to use SSL enabled port for RPC when its for node user.
-            establishRpc(messagingAddress, configuration).flatMap { rpc ->
+            establishRpc(p2pAddress, configuration).flatMap { rpc ->
                 rpc.waitUntilRegisteredWithNetworkMap().map {
                     NodeHandle(rpc.nodeIdentity(), rpc, configuration, process)
                 }
@@ -468,7 +468,7 @@ open class DriverDSL(
                         // TODO: remove the webAddress as NMS doesn't need to run a web server. This will cause all
                         //       node port numbers to be shifted, so all demos and docs need to be updated accordingly.
                         "webAddress" to apiAddress,
-                        "messagingAddress" to networkMapAddress.toString(),
+                        "p2pAddress" to networkMapAddress.toString(),
                         "useTestClock" to useTestClock
                 )
         )
@@ -539,7 +539,7 @@ open class DriverDSL(
             val process = builder.start()
             // TODO There is a race condition here. Even though the messaging address is bound it may be the case that
             // the handlers for the advertised services are not yet registered. Needs rethinking.
-            return addressMustBeBound(executorService, nodeConf.messagingAddress).map { process }
+            return addressMustBeBound(executorService, nodeConf.p2pAddress).map { process }
         }
 
         private fun startWebserver(
@@ -557,7 +557,7 @@ open class DriverDSL(
                 emptyList()
 
             val javaArgs = listOf(path) +
-                    listOf("-Dname=node-${nodeConf.messagingAddress}-webserver") + debugPortArg +
+                    listOf("-Dname=node-${nodeConf.p2pAddress}-webserver") + debugPortArg +
                     listOf(
                             "-cp", classpath, className,
                             "--base-directory", nodeConf.baseDirectory.toString())