+++ node/src/main/kotlin/net/corda/node/internal/Node.kt
@@ -142,8 +142,8 @@ class Node(override val configuration: FullNodeConfiguration,
 
     private fun makeLocalMessageBroker(): HostAndPort {
         with(configuration) {
-            val useHost = tryDetectIfNotPublicHost(messagingAddress.hostText)
-            val useAddress = useHost?.let { HostAndPort.fromParts(it, messagingAddress.port) } ?: messagingAddress
+            val useHost = tryDetectIfNotPublicHost(p2pAddress.hostText)
+            val useAddress = useHost?.let { HostAndPort.fromParts(it, p2pAddress.port) } ?: p2pAddress
             messageBroker = ArtemisMessagingServer(this, useAddress, rpcAddress, services.networkMapCache, userService)
             return useAddress
         }