+++ docs/source/release-notes.rst
@@ -10,8 +10,8 @@ Milestone 10
 Important: There are configuration changes in M10 due to the split of the Artemis port into separate messaging and RPC
 ports. To upgrade, you *must*:
 
- 1. In Gradle build configurations replace any references to ``artemisPort`` with ``messagingPort``.
- 2. In node configurations replace ``artemisAddress`` with ``messagingAddress``.
+ 1. In Gradle build configurations replace any references to ``artemisPort`` with ``p2pPort``.
+ 2. In node configurations replace ``artemisAddress`` with ``p2pAddress``.
 
 Milestone 9
 -----------